pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "world-model"
version = "0.1.0"
description = "Phase 1: Grid World + Predictive Model (Ha & Schmidhuber World Models)"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.1",
    "numpy>=1.24",
    "pygame>=2.5",
    "matplotlib>=3.7",
    "cma>=3.3",
    "pyyaml>=6.0",
    "tqdm>=4.65",
    "scikit-learn>=1.3",
]

[project.optional-dependencies]
dev = ["pytest>=7.4", "jupyter>=1.0"]

[tool.setuptools.packages.find]
include = ["world_model*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
